fix(net): stage 4a repair R6 - real SDK enrollment, not a raw-outcome approximation
Two independent blockers made the promised exchange impossible
through the intended API, both hidden by the raw core fixture.

1. **Codec composition.** `serve_enrollment(_auto)` registered a
   typed JSON handler returning `Vec<u8>`, and `TypedRpcHandler`
   encodes the return value - so the body reaching the core was a
   JSON array of numbers, while core promotion reads the body's own
   raw `NMO1` prefix. A legitimate `JoinOutcome::Admitted` could
   never promote. New `serve_rpc_raw_bytes` / `call_raw_bytes` pass
   bodies through unchanged, and enrollment + renewal use them on
   both sides: these bodies are already a serialized protocol.
2. **Reply subscription.** The origin-bound reply channel requires a
   TOFU-pinned entity, which a genuinely new provisional peer cannot
   have - gate 4 refuses to ingest its announcement. The binding now
   accepts a **bootstrap substitute** under three simultaneous
   conditions: the session is provisional, the channel is exactly
   that origin's enrollment reply channel, and the origin is the one
   the session already bound (one session, one identity). Nothing
   else changes: every other origin-bound channel still demands a
   pinned identity, which is the failure mode Kyra warned against.

Also: the rejecting fixture wrote its outcome code as `u32` where
the wire type is `u16` (prefix-only assertions missed it).

To exercise the real API the SDK gains `MeshBuilder::rtc(..)` behind
a new `webrtc` passthrough feature and a public `Mesh::node()`: a
browser-facing anchor must be constructible from the SDK, or the
service whose admission §12 gates cannot be tested on the surface
that serves it.

Witness `sdk/tests/enrollment_over_rtc.rs`: the unchanged intended
service and client, real signed `JoinRequest`, real single-use
invite, over the native RTC stand-in - the body is asserted to carry
raw `NMO1`, `Admitted` promotes the exact provisional incarnation on
its own endpoint, and an invite minted for another root is
`Rejected` and leaves the session provisional.

Inverses: restore the typed JSON registration -> RED; drop the
bootstrap origin binding -> RED. SDK units 292/292.

+++ net/crates/net/tests/rtc_admission.rs
@@ -498,7 +498,10 @@ fn outcome_bytes(admitted: bool) -> Bytes {
         buf.extend_from_slice(chain);
     } else {
         buf.push(1);
-        buf.extend_from_slice(&7u32.to_le_bytes());
+        // R6: the outcome CODE is a `u16` on the wire
+        // (`sdk/src/enrollment.rs`); the fixture used to write a
+        // `u32`, and prefix-only assertions never noticed.
+        buf.extend_from_slice(&7u16.to_le_bytes());
         let msg = b"invite expired";
         buf.extend_from_slice(&(msg.len() as u32).to_le_bytes());
         buf.extend_from_slice(msg);
